refactor(solver): shard mapped keyof constraint evaluation
